feat: Added multiple images panning together.
Did this (at least in theory) by moving the container instead of the image, and then just having the images scroll with the container. For some reason this bugged the top boundary so we lost that one.

Also added .env to the dockerignore so Gemini can use it's key as expected.